+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 Full documentation for rocSOLVER is available at [rocsolver.readthedocs.io](https://rocsolver.readthedocs.io/en/latest/).
 
 ## (Unreleased) rocSOLVER
+- LU factorization without pivoting for block tridiagonal matrices:
+    - GEBLTTRF_NPVT now has an interleaved\_batched version
+- Linear system solver without pivoting for block tridiagonal matrices:
+    - GEBLTTRS_NPVT now has an interleaved\_batched version
+
 ### Added
 ### Optimized
 ### Changed
@@ -20,7 +25,7 @@ Full documentation for rocSOLVER is available at [rocsolver.readthedocs.io](http
 - LU factorization without pivoting for block tridiagonal matrices:
     - GEBLTTRF_NPVT (with batched and strided\_batched versions)
 - Linear system solver without pivoting for block tridiagonal matrices:
-    - GEBLTTRS_NPVT (with batched and strided\_batched, versions)
+    - GEBLTTRS_NPVT (with batched and strided\_batched versions)
 - Product of triangular matrices
     - LAUUM
 - Added experimental hipGraph support for rocSOLVER functions